==> scripts/1_climate_indices.Rmd <==
---
title: "Generate Climate Indices"
author: "Inigo Peng"
date: "`r Sys.Date()`"
output: 
  github_document:
    toc: true
    toc_depth: 3
    number_sections: false

---

```{r setup, include=FALSE}
library(httr)
library(readr)
library(dplyr)
library(tidyverse)
library(lubridate)
library(zoo)
```

This R Markdown script is used to generate climate index inputs for NWI. 
Inputs: 

* PDO N34 min max (data/climate_indicies_data/min_max_pdo_n34.csv)

```{r}
normalizing_values <- read_csv(here::here( "data/climate_indicies_data/min_max_pdo_n34.csv"))
three_mta_norm_values <- read_csv(here::here("data/climate_indicies_data/normalizedsheet.csv"))
```
Outputs: 

* data-raw/n34.txt
* data-raw/pdo.txt
* data/climate_indicies_data/normalized_climate_indicies.csv

TODOs: None

### Download N34
```{r}
nino_url <- "https://psl.noaa.gov/data/correlation/nina34.anom.data"
workspace <- here::here("data-raw")
nino_file <- paste0(workspace,"/n34.txt")
download.file(nino_url, destfile = nino_file, mode = "wb")
```
### Read in and clean N34
```{r}
nino_raw <- read_lines(nino_file)
#Remove header
n34 <- nino_raw[-1]
# Convert to dataframe
n34_df <- read.table(text = n34, header = FALSE, fill = TRUE)
colnames(n34_df) <- c("Year", "Jan", "Feb", "Mar", "Apr", "May", "Jun",
                      "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")
n34_df_filtered <- n34_df|>
  filter(Year <= year(Sys.Date()))|>
  filter(Year > 1978)|>
  mutate(across(-Year, as.numeric))
clean_n34_df <- n34_df_filtered |> 
  pivot_longer(cols = -Year, names_to = "month", values_to = "N34")|>
  mutate(N34 = ifelse(N34 == -99.99, NA, N34))|>
  filter(!is.na(N34)) |> 
  mutate(
    month = match(tolower(month), tolower(month.abb)), 
    month = sprintf("%02d", month), 
    index = paste(Year, month, sep = "-"),
    month = as.numeric(month),
    Year = as.numeric(Year))|>
  relocate(index, .before = Year)
```
### Download PDO
```{r}
pdo_url <- "https://www.ncei.noaa.gov/pub/data/cmb/ersst/v5/index/ersst.v5.pdo.dat"
workspace <- here::here("data-raw")
pdo_file <- paste0(workspace, "/pdo.txt")
download.file(pdo_url, destfile = pdo_file, mode = "wb")
```
### Read in and clean PDO
```{r}
pdo_raw <- read_lines(pdo_file)
pdo<- pdo_raw[-1]
pdo_df <- read.table(text = pdo, header = FALSE, fill = TRUE)
colnames(pdo_df) <- c("Year", "Jan", "Feb", "Mar", "Apr", "May", "Jun",
                      "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")
pdo_df_filtered <- pdo_df|>
  filter(Year <= year(Sys.Date()))|>
  filter(Year > 1978)|>
  mutate(across(-Year, as.numeric))

clean_pdo_df <- pdo_df_filtered |> 
  pivot_longer(cols = -Year, names_to = "month", values_to = "PDO")|>
  mutate(PDO = ifelse(PDO == 99.99, NA, PDO))|>
  filter(!is.na(PDO)) |> 
  mutate(
    month = match(tolower(month), tolower(month.abb)), 
    month = sprintf("%02d", month),  # Ensure two-digit format
    index = paste(Year, month, sep = "-"),
    month = as.numeric(month),
    Year = as.numeric(Year))|>
  relocate(index, .before = Year)
```

### Normalize N34, PDO, generate three day moving average
```{r}
normalize <- function(x, min_val, max_val) {
  norm_value <- (x - min_val) / (max_val - min_val)
  return(max(0, min(1, norm_value)))
}

colnames(normalizing_values) <- tolower(colnames(normalizing_values))

# df_date_lag <- clean_n34_df |>
#   mutate(date = as.Date(paste0(index, "-01")) %m+% months(1)) |> 
#   dplyr::select(date)

clean_pdo_df_date <- clean_pdo_df |> 
  mutate(date = as.Date(paste0(index, "-01"))) |> glimpse()

max_date <- clean_pdo_df_date |> 
  pull(date) |> 
  max() |>            
  `%m+%`(months(1))

full_normalized_df <- clean_n34_df|>
  mutate(date = as.Date(paste0(index, "-01"))) |>
  # full_join(df_date_lag) |>
  full_join(clean_pdo_df_date, by = c("month", "index", "Year","date")) |>
  left_join(normalizing_values, by = "month")|>
  left_join(three_mta_norm_values, by="month") |> #TODO ADD ROW
  add_row(date=max_date) |>
  rowwise()|>
  mutate(normalized_CN34 = 1 - normalize(N34, minimum_of_n34, maximum_of_n34),
         normalized_pdo = normalize(PDO, minimum_of_pdo, maximum_of_pdo),
         combined_pdo_cn34 = normalized_CN34 + normalized_pdo,
         normalized_pdo_cn34 = normalize(combined_pdo_cn34, minimum_of_pdo_cn34, maximum_of_pdo_cn34)
         ) |> 
  ungroup() |> 
  mutate(
    three_mta_normalized_PDO = rollmean(normalized_pdo, k = 3, fill=NA, align = "right"),
    three_mta_normalized_PDO_CN34 = rollmean(normalized_pdo_cn34, k = 3, fill = NA, align = "right")) |>
  rowwise() |>
  mutate(
    normalized_3mta_normalized_PDO = normalize(three_mta_normalized_PDO, min_3mta_norm_PDO, max_3mta_norm_PDO),
    normalized_3mta_normalized_PDO_CN34 = normalize(three_mta_normalized_PDO_CN34, min_3mta_norm_PDO_CN34, max_3mta_norm_PDO_CN34)) |>  
  ungroup() |>  
  mutate(
    normalized_3mta_normalized_PDO_lag_0 = lag(normalized_3mta_normalized_PDO, n=1),
    normalized_3mta_normalized_PDO_CN34_lag_0 = lag(normalized_3mta_normalized_PDO_CN34, n=1)
  ) |> 
  rename(year=Year) |> glimpse()

full_normalized_df[nrow(full_normalized_df),]$normalized_3mta_normalized_PDO_CN34_lag_0 <- if_else(is.na(full_normalized_df[nrow(full_normalized_df)-1,]$normalized_3mta_normalized_PDO_CN34), full_normalized_df[nrow(full_normalized_df)-1,]$normalized_3mta_normalized_PDO_CN34_lag_0, full_normalized_df[nrow(full_normalized_df)-1,]$normalized_3mta_normalized_PDO_CN34)

cleaned_normalized_climate_indicies <- full_normalized_df[,c("date",
                                                          "year",
                                                          "month",
                                                          "PDO",
                                                          "normalized_pdo",
                                                          "N34",
                                                          "normalized_CN34",
                                                          "combined_pdo_cn34",
                                                          "normalized_pdo_cn34",
                                                          "three_mta_normalized_PDO",
                                                          "three_mta_normalized_PDO_CN34",   
"normalized_3mta_normalized_PDO",
"normalized_3mta_normalized_PDO_CN34",
"normalized_3mta_normalized_PDO_lag_0",
"normalized_3mta_normalized_PDO_CN34_lag_0")]

write.csv(cleaned_normalized_climate_indicies, here::here("data/climate_indicies_data/normalized_climate_indicies.csv"), row.names = FALSE)
```
