==> scripts/2_swe.Rmd <==
---
title: "Generate SWE"
author: "Inigo Peng"
date: "`r Sys.Date()`"
output: 
  github_document:
    toc: true
    toc_depth: 3
    number_sections: false
---

```{r setup, include=FALSE}
library(httr)
library(jsonlite)
library(lubridate)
library(tidyverse)
library(lubridate)
library(zoo)
library(dplyr)
```

This R Markdown script is used to generate SWE inputs for NWI. 
Inputs: 

* HUC 8 Station Codes  (data/swe_data/station_triplets)
* HUC 8 Area Proportions (data/swe_data/snotel-area.csv)
* SWE weighted averages  (data/swe_data/weighted_avg_swe.csv)

```{r}
station_triplets_lookup <- readRDS(here::here("data/swe_data/station_triplets")) |> glimpse()
snotel_area <- read_csv(here::here("data/swe_data/snotel-area.csv")) |> glimpse()
min_max_weight_swe <- read_csv(here::here("data/swe_data/weighted_avg_swe.csv")) |> glimpse() 
```

Outputs: 

* SWE Download (data/swe_data/swe_download.csv)
* SWE indices with values for NWI caluclation (data/swe_data/swe_summary.csv)

TODOs:
Decide on start date and end date for SWE time series data. Default start date is Oct 1 of the current water year. Default end date is current date. 
```{r}
timeseries_start_date = paste0(
  if_else(month(Sys.Date()) <=10, year(Sys.Date()) - 1, year(Sys.Date())), "-09-14")
timeseries_end_date = today()
```

### Download SWE Data
```{r}
# Function to get timeseries data
# Default element_cd is WTEQ (Snow Water Equivalent)
# Default start date is Oct 1 of the current water year
# Default end date is current date
# Default data temporal resolution is daily
get_station_data <- function(station_triplet){
  station_data <- get_timeseries_data(station_triplet = station_triplet)
  station_data_df <- as.data.frame(station_data[,2][[1]])$values[[1]]
  col_name <- station_triplets_lookup[station_triplet]
  colnames(station_data_df) <- c("Date", col_name)
  return(station_data_df)
}
get_timeseries_data <- function(station_triplet,
                                element_cd = "WTEQ",
                                temp_duration = "Daily",
                                start_date = timeseries_start_date,
                                end_date = timeseries_end_date) {
  url <- paste0(nrcs_api_url, "/services/v1/data")
  params <- list(
    stationTriplets = station_triplet,
    elements = element_cd,
    duration = temp_duration,
    beginDate = start_date,
    endDate = end_date,
    periodRef = "START"
  )
  response <- GET(url, query = params)
  if (status_code(response) == 200) {
    data <- content(response, "text", encoding = "UTF-8")
    return(fromJSON(data))
  } else {
    print(paste("Error:", status_code(response)))
    return(NULL)
  }
}
```
```{r}
nrcs_api_url <- "https://wcc.sc.egov.usda.gov/awdbRestApi"
data_list <- list()
for (station in names(station_triplets_lookup)) {
  station_df <- get_station_data(station)
  data_list[[station]] <- station_df
}
full_swe_df <- Reduce(function(x, y) full_join(x, y, by = "Date"), data_list) |> glimpse()
write_csv(full_swe_df, here::here("data/swe_data/swe_download.csv"))
```

### Calculate basin averages, weighted averaged, and normalized values
```{r}
normalize <- function(x, min_val, max_val) {
  if (isTRUE(all(c(x, min_val, max_val) == 0, na.rm = TRUE))) {
    return(0)
  }
  norm_value <- (x - min_val) / (max_val - min_val)
  return(max(0, min(1, norm_value)))  # Ensure within [0,1] range
}

swe_summary <- full_swe_df |> 
  mutate(
    water_year = ifelse(month(Date) >= 10, year(Date) + 1, year(Date))) |> 
  group_by(water_year) |>  
  mutate(
    day_of_water_year = as.integer(difftime(Date, ymd(paste0(water_year - 1 ,'-09-30')), units = "days"))) |> 
  ungroup() |> 
  mutate(
    ukl_huc8_mean_swe = lag(rowMeans(pick(
      `Billie Creek Divide SWE (in)`,
      `Cold Springs Camp SWE (in)`,
      `Fish Lake SWE (in)`,
      `Fourmile Lake SWE (in)`,
      `Sevenmile Marsh SWE (in)`
      )
  ), n=1),
  williamson_huc8_mean_swe = lag(rowMeans(pick(
    `Diamond Lake SWE (in)`,
    `Chemult Alternate SWE (in)`,
    `Silver Creek SWE (in)`, 
    `Taylor Butte SWE (in)`
    )
  ), n=1),
  sprague_huc8_mean_swe = lag(rowMeans(pick(
    `Silver Creek SWE (in)`,
    `Taylor Butte SWE (in)`,
    `Summer Rim SWE (in)`,
    `Quartz Mountain SWE (in)`,
    `Strawberry SWE (in)`
    )
  ), n=1)) |> 
  mutate_all(funs(replace_na(.,0))) |> 
  mutate(weighted_average_swe = 
           (ukl_huc8_mean_swe*snotel_area$`Proportion of area >1500 m`[1]) + 
           (williamson_huc8_mean_swe*snotel_area$`Proportion of area >1500 m`[2]) + 
           (sprague_huc8_mean_swe*snotel_area$`Proportion of area >1500 m`[3])) |> 
  left_join(min_max_weight_swe, by = "day_of_water_year") |>
  rowwise()|>
  mutate(normalized_average_swe = normalize(weighted_average_swe, 
                                                  min_weighted_average_swe_in, 
                                                  max_weighted_average_swe_in))|>
  ungroup() |>
  group_by(water_year) |> 
  mutate(
    swe_may15 = normalized_average_swe[
      month(Date) == 5 & day(Date) == 15
    ][1],
    normalized_average_swe = case_when(
      day_of_water_year == 366 ~ lag(normalized_average_swe),
      Date >= ymd(paste0(water_year - 1, "-05-16")) &
      Date <= ymd(paste0(water_year - 1, "-06-30")) ~ swe_may15,
      TRUE ~ normalized_average_swe
    )
  ) |> 
  mutate(
    day_after_oct_15 = as.integer(difftime(Date, ymd(paste0(year(Date), "-10-15")), units = "days")),
    day_after_oct_15 = ifelse(Date > ymd(paste0(year(Date), "-11-30")) | day_after_oct_15 < 1, NA, day_after_oct_15),
    oct_16_nov_30_linear_weight = day_after_oct_15 * (1/46),
    day_after_may_15_desc = 47 - as.integer(difftime(Date, ymd(paste0(year(Date), "-5-15")), units = "days")),
    day_after_may_15_desc = ifelse(Date > ymd(paste0(year(Date), "-6-30")) | Date < ymd(paste0(year(Date), "-5-16")), NA, day_after_may_15_desc),
    may_16_jun_30_linear_weight = day_after_may_15_desc * (1/46),
    normalized_weighted_avg_swe = case_when(!is.na(day_after_oct_15) ~ normalized_average_swe * oct_16_nov_30_linear_weight,
                                            !is.na(day_after_may_15_desc) ~ swe_may15 * may_16_jun_30_linear_weight,
                                            .default = normalized_average_swe)) |> 
  dplyr::select(-c(min_weighted_average_swe_in, 
            max_weighted_average_swe_in, 
            max_weighted_normalized_average_swe, 
            min_weighted_normalized_average_swe)) |> 
  relocate(water_year, .after = Date) |> 
  relocate(day_of_water_year, .after = water_year) |> 
  rename(date = Date) |> 
  mutate(date = as.Date(date, format= "%Y-%m-%d")) |> 
  dplyr::select(-c(
    `Billie Creek Divide SWE (in)`,
    `Cold Springs Camp SWE (in)`,
    `Fish Lake SWE (in)`,
    `Fourmile Lake SWE (in)`,
    `Sevenmile Marsh SWE (in)`,
    `Diamond Lake SWE (in)`,
    `Chemult Alternate SWE (in)`,
    `Silver Creek SWE (in)`, 
    `Taylor Butte SWE (in)`,
    `Summer Rim SWE (in)`,
    `Quartz Mountain SWE (in)`,
    `Strawberry SWE (in)`,
    swe_may15
      ))
  

write.csv(swe_summary, here::here("data/swe_data/swe_summary.csv"), row.names = FALSE)
```
