==> scripts/3_prism.Rmd <==
---
title: "Generate Prism Data"
author: "Inigo Peng"
date: "`r Sys.Date()`"
output: 
  github_document:
    toc: true
    toc_depth: 3
    number_sections: false
---

```{r setup, include=FALSE}
library(httr)
library(readr)
library(ncdf4)
library(stringr)
library(tidyverse)
library(raster)
library(zoo)
```

This R Markdown script is used to generate Prism inputs for NWI. 
Inputs: 

* Prism Locations with Coordinates  (data/prism_data/prism_locations) - generated through lookups.R
* HUC 8 Area Proportions (data/swe_data/snotel-area.csv)
* Precipitation min max  (data/prism_data/min_max_ppt.csv)
* Downloaded precipitation historic data starting Oct 2020 (data/prism_data/historic_prism_ppt_download.csv)


```{r}
prism_locations <- readRDS(here::here("data/prism_data/prism_locations")) |> glimpse()
snotel_area <- read_csv(here::here("data/swe_data/snotel-area.csv")) |> glimpse()
min_max_ppt <- read_csv(here::here("data/prism_data/min_max_ppt.csv")) |> glimpse()
# ppt_long_historic_data <- read_csv(here::here("data/prism_data/historic_prism_ppt_download.csv")) |> glimpse()
ppt_hist_early <- read_csv(here::here("data/prism_data/early_hist_prism_ppt.csv")) |>
  mutate(date = as.Date(date, format = "%m/%d/%Y"))
```

Outputs: 

* PPT Download (data/prism_data/recent_prism_ppt_download.csv)
* PPT, Tmean, Tmax joined download (data/prism_data/prism_ppt_tmean_tmax_download.csv)
* PPT lag and normalized values for NWI input (data/prism_data/prism_ppt_summary_export.csv)
* PPT, Tmean, Tmax summarized data (data/prism_data/prism_summary.csv)

TODOs:
* Decide on start date and end date for Prism time series data. Default start date is 16 days before current date. Default end date is one day before current date.
* Decide on download location director - downloading historical files could add up so make sure there's enough disk space. If choosing a separate directory, make sure its structured as ".../Prism_Files/{element}/
* If downloading historic data, only need PPT.

Note: also first time generation will need historic data (back to Oct 2020) to generate the values for `trailing_sum_31_1095_d_precip_in`. This takes a long time to download and process 4 + years of data. I have included the downloaded and extracted data from 2020 until  

```{r}
last_date_recorded <- max(ppt_hist_early$date)
# prism_start_date = Sys.Date() - 16
prism_start_date = last_date_recorded + 1
prism_end_date = Sys.Date() - 1
if (prism_start_date > prism_end_date){
  warning("Start date before end date")
}
# prism_end_date = as.Date("2025-3-30")
# element_list <- c("ppt", "tmean", "tmax")
element_list <- c("ppt")
# local_dir <- "Z:/Shared/Active_Projects/083-02_SDM_Support_Base_Year/Modeling/NWI"
local_dir <- here::here("data-raw/")
```

### Prism Data Download
```{r}
normalize <- function(x, min_val, max_val) {
  norm_value <- (x - min_val) / (max_val - min_val)
  return(max(0, min(1, norm_value)))  # Ensure within [0,1] range
}
prism_base_url <- "http://services.nacse.org/prism/data/public/4km"
prism_coordinates <- apply(prism_locations[,2:3], 2, as.numeric)
prism_ppt_data <- as.data.frame(prism_locations)
prism_tmean_data <- as.data.frame(prism_locations)
prism_tmax_data <- as.data.frame(prism_locations)
```

### Loop through elements, download zipped nc files, unzip nc files and parsel for the interested location using coordinates
```{r}
for (element in element_list){
  for (current_date in seq(prism_start_date, prism_end_date, by = "day")) {
    day <- format(as.Date(current_date), "%Y%m%d")
    download_url <- paste0(prism_base_url, "/", element, "/", day, "?format=nc")
    download.file(download_url, destfile = paste0(local_dir,"/Prism_Files/",element,"/",day,"_",element, ".zip"), mode = "wb")
    Sys.sleep(2)
  }
}
```
```{r}
for (element in element_list){
  #TODO errors out on unzipped folder
  # nc_files <- list.files(path = paste0(local_dir,"/Prism_Files/", element), pattern = "\\.zip$", full.names = TRUE)
  nc_files <- c()
  for (date in seq(prism_start_date, prism_end_date, by = "day")) {
    ds <- format(as.Date(date), "%Y%m%d")
    daily_files <- list.files(
      path       = paste0(local_dir,"/Prism_Files/", element),
      pattern    = paste0(ds, ".*\\.zip$"), # starts with date, ends in .zip
      full.names = TRUE
    )
    # accumulate:
    nc_files <- c(nc_files, daily_files)
}
  for (file in nc_files){
    print(file)
    date_str <- str_extract(file, "[0-9]{8}")
    unzip(file, exdir = paste0(local_dir, "/Prism_Files/", element, "/unzipped_files"))
    base_path <- paste0(local_dir,"/Prism_Files/", element, "/unzipped_files/PRISM_", element, "_")
    possible_files <- c(
      paste0(base_path, "provisional_4kmD2_", date_str, "_nc.nc"),
      paste0(base_path, "stable_4kmD2_", date_str, "_nc.nc"),
      paste0(base_path, "early_4kmD2_", date_str, "_nc.nc")
    )
    nc_file <- possible_files[file.exists(possible_files)][1]
    print(nc_file)
    ncin <- nc_open(nc_file)
    lon <- ncvar_get(ncin,"lon")
    lat <- ncvar_get(ncin, "lat")
    element_val <-ncvar_get(ncin, "Band1")
    fillvalue <- ncatt_get(ncin, "Band1", "_FillValue")
    nc_close(ncin)
    element_val[element_val == fillvalue$value] <- NA
    r <- raster::raster(t(element_val), xmn=min(lon), xmx=max(lon), ymn=min(lat), ymx=max(lat), crs=CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs+ towgs84=0,0,0"))
    r <- flip(r, direction='y')
    date_obj <- as.Date(date_str, format="%Y%m%d")
    raster_data <- raster::extract(r, cbind(prism_coordinates[,"longitude"],prism_coordinates[,"latitude"]), method='simple')
    if (element == "ppt"){
      ppt_inch <- raster_data/25.4
      ppt_inch <- as.data.frame(ppt_inch)
      colnames(ppt_inch) <- date_obj
      prism_ppt_data <- cbind(prism_ppt_data, ppt_inch)
    }else if(element == "tmean") {
      tmean_f <- (raster_data*9/5) + 32
      tmean_f <- as.data.frame(tmean_f)
      colnames(tmean_f) <- date_obj
      prism_tmean_data <- cbind(prism_tmean_data, tmean_f)
    }else{
      tmax_f <- (raster_data*9/5) + 32
      tmax_f <- as.data.frame(tmax_f)
      colnames(tmax_f) <- date_obj
      prism_tmax_data <- cbind(prism_tmax_data, tmax_f)
    }
  }
}
```

### Download extract PPT data
```{r}
ppt_long_recent_data <- prism_ppt_data |>
  dplyr::select(-c(longitude, latitude)) |>
  pivot_longer(!name, names_to = "date", values_to = "ppt") |> 
  mutate(date = as.Date(date)) |> glimpse()
write_csv(ppt_long_recent_data, here::here("data/prism_data/early_prism_ppt_download.csv"))

ppt_long_data_early <- ppt_long_recent_data |>
  mutate(group = case_when(
    startsWith(name, "UKL") ~ "mean_total_daily_precip_ukl",
    startsWith(name, "WIL") ~ "mean_total_daily_precip_williamson",
    startsWith(name, "SPR") ~ "mean_total_daily_precip_sprague"
  )) |>
  group_by(date, group) |>
  summarise(mean = mean(ppt)) |>
pivot_wider(names_from = "group", values_from ="mean")
ppt_long_data_early <- bind_rows(ppt_hist_early, ppt_long_data_early) 
write_csv(ppt_long_data_early, here::here("data/prism_data/early_hist_prism_ppt.csv"))
```
```{r}
# ppt_long_recent_data <- prism_ppt_data |>
#   dplyr::select(-c(longitude, latitude)) |>
#   pivot_longer(!name, names_to = "date", values_to = "ppt") |> 
#   mutate(date = as.Date(date)) |> glimpse()
# write_csv(ppt_long_recent_data, here::here("data/prism_data/provisional_recent_prism_ppt.csv"))
# 
# ppt_hist_provisional <- read_csv(here::here("data/prism_data/provisional_hist_prism_ppt.csv")) |>
#   mutate(date = as.Date(date, format = "%m/%d/%Y")) |>
#   filter(date < "2025-03-15") |> 
#   mutate(group = case_when(
#     startsWith(name, "UKL") ~ "mean_total_daily_precip_ukl",
#     startsWith(name, "WIL") ~ "mean_total_daily_precip_williamson",
#     startsWith(name, "SPR") ~ "mean_total_daily_precip_sprague"
#   )) |>
#   group_by(date, group) |>
#   summarise(mean = mean(ppt)) |>
#   pivot_wider(names_from = "group", values_from ="mean")

```
```{r}
# ppt_long_data_provisional <- ppt_long_recent_data |>
#   mutate(group = case_when(
#     startsWith(name, "UKL") ~ "mean_total_daily_precip_ukl",
#     startsWith(name, "WIL") ~ "mean_total_daily_precip_williamson",
#     startsWith(name, "SPR") ~ "mean_total_daily_precip_sprague"
#   )) |>
#   group_by(date, group) |>
#   summarise(mean = mean(ppt)) |>
# pivot_wider(names_from = "group", values_from ="mean")
# ppt_long_data_provisional <- bind_rows(ppt_hist_provisional, ppt_long_data_provisional) |> 
#   arrange(date)

```
### Bind historic data with recent data
Recent data value of the same date replaces historic data values
```{r}
# ppt_recent_early <- read_csv(here::here("data/prism_data/early_recent_prism_ppt.csv")) |> 
#   pivot_longer(!Date, names_to = "name", values_to = "ppt") |> 
#   mutate(date = as.Date(Date, format = "%m/%d/%Y")) |> 
#   dplyr::select(-("Date")) |> 
#   mutate(group = case_when(
#     startsWith(name, "UKL") ~ "mean_total_daily_precip_ukl",
#     startsWith(name, "WIL") ~ "mean_total_daily_precip_williamson",
#     startsWith(name, "SPR") ~ "mean_total_daily_precip_sprague"
#   )) |>
#   group_by(date, group) |>
#   summarise(mean = mean(ppt)) |>
#   pivot_wider(names_from = "group", values_from ="mean")
```
```{r}
# ppt_long_data_early <- bind_rows(ppt_hist_early, ppt_early) |> 
#   arrange(date)
```
### Generate data PPT data summaries, normalized values 
```{r}
ppt_data_summary <- ppt_long_data_early |>
# ppt_data_summary <- ppt_long_data_provisional |>
  mutate(water_year = ifelse(month(date) >= 10, year(date) + 1, year(date))) |>
  group_by(water_year) %>%
  mutate(
    day_of_water_year = as.integer(difftime(date, ymd(paste0(water_year - 1 ,'-09-30')), units = "days"))) |>
  ungroup() |>
  arrange(date) |>
  mutate(
    mean_total_daily_precip_ukl = lead(mean_total_daily_precip_ukl, 1),
    mean_total_daily_precip_williamson = lead(mean_total_daily_precip_williamson, 1),
    mean_total_daily_precip_sprague = lead(mean_total_daily_precip_sprague, 1)
  ) |>
  relocate("mean_total_daily_precip_sprague", .after = "mean_total_daily_precip_williamson") |>
  relocate(c("water_year", "day_of_water_year"), .after = "date") |>
  mutate(weighted_mean_total_daily_precipitation_ukl_catchment_in = lag(
         (mean_total_daily_precip_ukl*snotel_area$`Proportion of total area`[1] +
         mean_total_daily_precip_williamson*snotel_area$`Proportion of total area`[2]+
         mean_total_daily_precip_sprague*snotel_area$`Proportion of total area`[3]), 1, default = 0),
         thirty_d_trailing_sum_precip = lag(rollsum(weighted_mean_total_daily_precipitation_ukl_catchment_in, k = 30, fill=NA, align = "right"))
         ) |>
  left_join(min_max_ppt, by = "day_of_water_year") |>
  rowwise()|>
  mutate(normalized_30_trailing_sum_precipitation = normalize(thirty_d_trailing_sum_precip,
                                                                    min_30_trailing_sum_ppt,
                                                                    max_30_trailing_sum_ppt))|>
  ungroup() |>
  mutate(
    normalized_30_trailing_sum_precipitation = if_else(
      day_of_water_year == 366, lag(normalized_30_trailing_sum_precipitation),
      normalized_30_trailing_sum_precipitation
    )
  ) |> 
  mutate(date = as.Date(date, format= "%Y-%m-%d"))


ppt_data_export <- ppt_data_summary |>
  arrange(date) |>
  rowwise() |>
  mutate(
    trailing_sum_31_1095_d_precip_in = purrr::map_dbl(date, function(d) {
      lag_end <- d %m-% months(1) - days(1)
      lag_start <- lag_end - days(1064 - 1)
      
      vals <- ppt_data_summary$weighted_mean_total_daily_precipitation_ukl_catchment_in[
        ppt_data_summary$date >= lag_start & ppt_data_summary$date <= lag_end
      ]
      
      if (length(vals) == 1064) {
        sum(vals, na.rm = TRUE)
      } else {
        NA_real_
      }
    })
  ) |> 
mutate(
    normalized_trailing_sum_31_1095_d_precip_in = normalize(trailing_sum_31_1095_d_precip_in,
                                                            min_31_1095_d_trailing_sum_ppt,
                                                            max_of_31_1095_d_trailing_sum_ppt)
  ) |>
  ungroup() |>
  mutate(
    normalized_trailing_sum_31_1095_d_precip_in = if_else(
      day_of_water_year == 366, lag(normalized_trailing_sum_31_1095_d_precip_in),
      normalized_trailing_sum_31_1095_d_precip_in
    )
  ) |> 
  dplyr::select(-c(min_30_trailing_sum_ppt, max_30_trailing_sum_ppt, max_of_31_1095_d_trailing_sum_ppt, min_31_1095_d_trailing_sum_ppt)) 

write_csv(ppt_data_export, here::here("data/prism_data/early_prism_ppt_summary_export.csv"))

# write_csv(ppt_data_export, here::here("data/prism_data/provisional_prism_ppt_summary_export.csv"))
```
## Only run these chunks if downloaded Tmean and Tmax Data

```{r}
tmean_long_data <- prism_tmean_data |>
  dplyr::select(-c(longitude, latitude)) |>
  pivot_longer(!name, names_to = "date", values_to = "tmean")
tmax_long_data <- prism_tmax_data |>
  dplyr::select(-c(longitude, latitude)) |>
  pivot_longer(!name, names_to = "date", values_to = "tmax")
full_download_data <- ppt_long_data |>
  full_join(tmean_long_data, by = c("name", "date")) |>
  full_join(tmax_long_data, by = c("name", "date"))
  # mutate(across(where(is.numeric), ~ round(.x, 3)))
write_csv(full_download_data, here::here("data/prism_data/prism_ppt_tmean_tmax_download.csv"))
```

```{r}
tmean_data_summary <- tmean_long_data |>
  mutate(group = case_when(
    startsWith(name, "UKL") ~ "UKL_tmean",
    startsWith(name, "WIL") ~ "Williamson_tmean",
    startsWith(name, "SPR") ~ "Sprague_tmean"
  )) |>
  group_by(date, group) |>
  summarise(mean = mean(tmean)) |>
  pivot_wider(names_from = "group", values_from ="mean")

tmax_data_summary <- tmax_long_data |>
  mutate(group = case_when(
    startsWith(name, "UKL") ~ "UKL_tmax",
    startsWith(name, "WIL") ~ "Williamson_tmax",
    startsWith(name, "SPR") ~ "Sprague_tmax"
  )) |>
  group_by(date, group) |>
  summarise(mean = mean(tmax)) |>
  pivot_wider(names_from = "group", values_from ="mean")

summary_data <- ppt_data_summary |>
  full_join(tmean_data_summary, by = c("date")) |>
  full_join(tmax_data_summary, by = c("date"))
  # mutate(across(where(is.numeric), ~ round(.x, 3)))
write_csv(summary_data, here::here("data/prism_data/prism_summary.csv"))
```
