==> scripts/5_nwi.Rmd <==
---
title: "5_Generate_NWI"
author: "Inigo Peng"
date: "`r Sys.Date()`"
output: 
  github_document:
    toc: true
    toc_depth: 3
    number_sections: false
---
```{r setup, include=FALSE}
library(tidyverse)
library(dplyr)
library(lubridate)
```

This R Markdown script is used to generate NWI using outputs from previous Rmd. 
Inputs: 

* NWI min max for normalization  (data/nwi_data/nwi_min_max.csv)
* Model weights (data/nwi_data/nwi_model_weights.csv)
* Prism data  (data/prism_data/prism_ppt_summary_export.csv)
* Climate indicies (data/climate_indicies_data/normalized_climate_indicies.csv)
* SWE (data/swe_data/swe_summary.csv)
* UKLNI net inflow (data/ukr_net_inflow/net_inflow_export.csv)
* Leap year period and non leap year period lookup (data/nwi_data) 

```{r}
normalize <- function(x, min_val, max_val) {
  if (isTRUE(all(c(x, min_val, max_val) == 0, na.rm = TRUE))) {
    return(0)
  }
  norm_value <- (x - min_val) / (max_val - min_val)
  return(max(0, min(1, norm_value)))  # Ensure within [0,1] range
}
nwi_min_max <- read_csv(here::here("data/nwi_data/nwi_min_max.csv")) |> glimpse()
model_weights <- read_csv(here::here("data/nwi_data/nwi_model_weights.csv")) |> 
  dplyr::select(-c("date", "day_of_water_year")) |> glimpse()
prism_data <- read_csv(here::here("data/prism_data/early_prism_ppt_summary_export.csv")) |> glimpse()
climate_indicies <- read_csv(here::here("data/climate_indicies_data/normalized_climate_indicies.csv")) |> glimpse()
swe <- read_csv(here::here("data/swe_data/swe_summary.csv")) |> glimpse()
uklni_net_inflow <- read_csv(here::here("data/ukr_net_inflow/net_inflow_export.csv")) |> glimpse()
leap_year_period <- read_csv(here::here("data/nwi_data/leap_period_lookup.csv")) |> glimpse()
non_leap_year_period <- read_csv(here::here("data/nwi_data/non_leap_period_lookup.csv")) |> glimpse()
```
Outputs: 

* NWI (data/nwi_data/nwi_cleaned.csv)

TODOs:
Default NWI start date is september 15 of previous water year.
```{r}
current_water_year <- ifelse(month(Sys.Date()) >= 10, as.integer(year(Sys.Date()) + 1), as.integer(year(Sys.Date())))
start_year <- current_water_year - 1
start_date <- as.Date(paste0(year(Sys.Date()) - 1, "-09-14"))
```

### Expand Climate Indicies
```{r}
expanded_climate_indicies <- climate_indicies |>
  # mutate(date=as.Date(date, format="%m/%d/%Y")) |>
  rowwise() %>%
  mutate(date = list(seq.Date(date, ceiling_date(date, "month") - days(1), by = "day"))) |>
  unnest(date) |> 
  mutate(year = as.integer(format(date, "%Y")),
         month = as.integer(format(date, "%m")))
```

### Calculate NWI, normalized NWI and 2 week lag NWI
$$
WI_d = q_d Q_{d-1} + s_d S_{d-1} + pn_d PN_{d-1} + pl_d PL_{d-1} + c_d C_{m-1}
$$
```{r}
prism_data_edit <-prism_data |>
  filter(water_year >= start_year & date > start_date) |> 
  mutate(month = as.integer(format(date, "%m")),
         day = as.integer(format(date, "%d")),
         year = as.integer(format(date, "%Y")),
         water_year = if_else(month >= 10, year + 1, year),
         leap_year_index = if_else(leap_year(water_year), 1, 0)
         )
prism_leap <- prism_data_edit |> filter(leap_year_index == 1) |> 
  left_join(leap_year_period, by = c("day_of_water_year")) |> 
  dplyr::select(-c("lookup_date", "lookup_month")) |> glimpse()
prism_non_leap <- prism_data_edit |> filter(leap_year_index == 0) |> 
  left_join(non_leap_year_period, by = c("day_of_water_year")) |> 
  dplyr::select(-c("lookup_date")) |> glimpse()

prism_full <- bind_rows(prism_leap, prism_non_leap) |> arrange(date)
nwi <- prism_full |> 
  left_join(swe, by = c("date", "water_year", "day_of_water_year")) |>
  left_join(uklni_net_inflow, by = c("date", "day_of_water_year", "water_year")) |>
  left_join(expanded_climate_indicies, by=c("date", "month", "year")) |>
  mutate(day_of_water_year = as.integer(day_of_water_year),
         normalized_weighted_avg_swe = if_else(day_of_water_year < 16 | (month >= 7 & month <=9), NA, normalized_weighted_avg_swe)
         ) |>
  left_join(model_weights, by = c("period")) |>
  mutate(weight_for_30_d_trailing_sum_UKLNI = day_of_water_year*wi_slope + wi_intercept,
         weight_for_30_d_trailing_sum_precip = day_of_water_year*wp_30_slope + wp_30_intercept,
         weight_for_tot_31_1095_d_trailing_sum_precip = day_of_water_year*wp_1095_slope + wp_1095_intercept,
         weight_for_normalized_weighted_avg_swe = day_of_water_year*ws_slope + ws_intercept,
         weight_for_climate_index = day_of_water_year*wc_slope + wc_intercept,
         climate_index_lag_0 = ifelse(climate_index_used == "3 mta PDO_CN34", normalized_3mta_normalized_PDO_CN34_lag_0, normalized_3mta_normalized_PDO_lag_0),
         wetness_index_v20 = 
           normalized_30_d_trailing_sum_UKLNI*weight_for_30_d_trailing_sum_UKLNI +
           normalized_30_trailing_sum_precipitation*weight_for_30_d_trailing_sum_precip +
           normalized_trailing_sum_31_1095_d_precip_in*weight_for_tot_31_1095_d_trailing_sum_precip +
           replace_na(normalized_weighted_avg_swe, 0)*weight_for_normalized_weighted_avg_swe +
           climate_index_lag_0*weight_for_climate_index
         ) |>
  left_join(nwi_min_max, by=c("day_of_water_year")) |>
  rowwise() |>
  mutate(normalized_wetness_index_v20 = normalize(wetness_index_v20, min_nwi_v20, max_nwi_v20)) |>
  ungroup() |>
  mutate(
    normalized_wetness_index_v20 = if_else(
      day_of_water_year == 366, lag(normalized_wetness_index_v20),
      normalized_wetness_index_v20
    )
  ) |> 
  mutate(trailing_2_week_mean_nwi_v20 = zoo::rollmean(normalized_wetness_index_v20, k=14, fill=NA, align = "right"))
nwi_cleaned <- nwi[,c("date", "month", "day", "year", "day_of_water_year", "leap_year_index",
       "period",
       "normalized_30_d_trailing_sum_UKLNI",
       "weight_for_30_d_trailing_sum_UKLNI",
       "normalized_30_trailing_sum_precipitation",
       "weight_for_30_d_trailing_sum_precip",
       "normalized_trailing_sum_31_1095_d_precip_in",
       "weight_for_tot_31_1095_d_trailing_sum_precip",
       "normalized_weighted_avg_swe",
       "weight_for_normalized_weighted_avg_swe",
       "normalized_3mta_normalized_PDO_lag_0",
       "normalized_3mta_normalized_PDO_CN34_lag_0",
       "climate_index_used",
       "climate_index_lag_0",
       "weight_for_climate_index",
       "wetness_index_v20",
       "normalized_wetness_index_v20",
       "trailing_2_week_mean_nwi_v20",
       "ukl_net_inflow_smoothed")]
write_csv(nwi_cleaned, here::here("data/nwi_data/nwi_export.csv"))
```
